+++ src/lib.rs
@@ -980,9 +980,7 @@ pub enum TocKind {
 
 impl fmt::Display for TocKind {
 	#[inline]
-	fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-		f.write_str(self.as_str())
-	}
+	fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result { f.pad(self.as_str()) }
 }
 
 impl TocKind {