+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Changed
 
 * Add (more) inline hints
+* Add `Formatter` width/fill/align/etc. support for `AccurateRip`, `Cddb`, `ShaB64`
 * Improved documentation